Change pos if out of screen

+++ start/plane.h
@@ -19,7 +19,7 @@ public:
 	/// @param deltaTime the elapsed time in seconds
 	/// @return void
 	virtual void update(float deltaTime);
-	int fuel;
+	float fuel;
 private:
 	/* add your private declarations */
 public:
